fixed some styling on login page
